+++ src/JsonTypes/Hashtag2.cs
@@ -5,6 +5,6 @@ namespace TwitterDataMiner.JsonTypes
     public class Hashtag2
     {
         public string text { get; set; }
-        public List<int> indices { get; set; }
+        public List<long> indices { get; set; }
     }
 }
