+++ src/JsonTypes/Medium2.cs
@@ -2,8 +2,8 @@ namespace TwitterDataMiner.JsonTypes
 {
     public class Medium2
     {
-        public int w { get; set; }
-        public int h { get; set; }
+        public long w { get; set; }
+        public long h { get; set; }
         public string resize { get; set; }
     }
 }
