+++ src/JsonTypes/SearchMetadata.cs
@@ -8,8 +8,8 @@ namespace TwitterDataMiner.JsonTypes
         public string next_results { get; set; }
         public string query { get; set; }
         public string refresh_url { get; set; }
-        public int count { get; set; }
-        public int since_id { get; set; }
+        public long count { get; set; }
+        public long since_id { get; set; }
         public string since_id_str { get; set; }
     }
 }
