+++ src/JsonTypes/Url2.cs
@@ -7,6 +7,6 @@ namespace TwitterDataMiner.JsonTypes
         public string url { get; set; }
         public string expanded_url { get; set; }
         public string display_url { get; set; }
-        public List<int> indices { get; set; }
+        public List<long> indices { get; set; }
     }
 }
